fix: reject non-numeric github-user-id input
Number.parseInt silently truncates non-numeric suffixes ("123abc" ->
123) and drops decimals ("12.5" -> 12). Switch to Number() so these
produce NaN, and add .int() to the Zod schema to reject floats.

+++ src/schemas.ts
@@ -14,7 +14,7 @@ const BaseInputsSchema = z.object({
 });
 
 const WithGithubUserIDSchema = BaseInputsSchema.extend({
-	githubUserID: z.number().min(1),
+	githubUserID: z.number().int().min(1),
 	coderUsername: z.undefined(),
 });
